feat: add entrypoint for rewriter

+++ src/engine/symlevel/io/stream_file_reader.h
@@ -17,13 +17,13 @@ namespace IO {
 class StreamFileReader {
 public:
 
-    StreamFileReader(RandomAccessFile& file, size_t position): file(file), position(position) {}
-    StreamFileReader(RandomAccessFile* file, size_t position): file(*file), position(position) {}
+    StreamFileReader(RandomAccessFile& file, uint32_t position): file(file), position(position) {}
+    StreamFileReader(RandomAccessFile* file, uint32_t position): file(*file), position(position) {}
 
     /**
      * @brief Returns current stream position.
      */
-    size_t Position() const
+    uint32_t Position() const
     {
         return position;
     }
@@ -64,7 +64,7 @@ public:
 
 private:
     RandomAccessFile& file;
-    size_t position;
+    uint32_t position;
 
     template <typename T> T ReadValue()
     {
